Are They Equal

+++ patA/patA1060.cpp
@@ -1,39 +1,39 @@
-#include<iostream>
-#include<string>
+#include <iostream>
+#include <string>
 using namespace std;
 int n;
 
-string deal(string s,int& e){
-    int k=0;
-    while(s.length()>0 && s[0]=='0'){
-        s.erase(s.begin());        
+string deal(string s, int &e) {
+    int k = 0;
+    while(s.length() > 0 && s[0] == '0') {
+        s.erase(s.begin());
     }
-    if(s[0]=='.'){
+    if(s[0] == '.') {
         s.erase(s.begin()); 
-        while(s.length()>0 && s[0]=='0'){
+        while(s.length() > 0 && s[0] == '0'){
             s.erase(s.begin());
-            e--;        
+            e--;
         }
-    }else{
-        while(k<s.length() && s[k]!='.'){
+    } else {
+        while(k < (int)s.length() && s[k] != '.') {
             k++;
             e++;
         }
-        if(k<s.length()){
-            s.erase(s.begin()+k);
+        if(k < (int)s.length()) {
+            s.erase(s.begin() + k);
         }
     }
-    if(s.length()==0){
-        e=0;
+    if(s.length() == 0) {
+        e = 0;
     }
-    int num=0;
-    k=0;
+    int num = 0;
+    k = 0;
     string res;
-    while(num<n){
-        if(k<s.length()){
-            res+=s[k++];
-        }else {
-            res+='0';
+    while(num < n) {
+        if(k < (int)s.length()) {
+            res += s[k++];
+        } else {
+            res += '0';
         }
         num++;
     }
@@ -42,16 +42,15 @@ string deal(string s,int& e){
 
 
 int main(){
-    string s1,s2,s3,s4;
-    cin>>n>>s1>>s2;
-    int e1=0,e2=0;
-    s3=deal(s1,e1);
-    s4=deal(s2,e2);
-    if(s3==s4 && e1==e2){
-        cout<<"YES 0."<<s3<<"*10^"<<e1<<endl;
-    }else{
-        cout<<"NO 0."<<s3<<"*10^"<<e1<<" 0."<<s4<<"*10^"<<e2<<endl;
+    string s1, s2, s3, s4;
+    cin >> n >> s1 >> s2;
+    int e1 = 0, e2 = 0;
+    s3 = deal(s1, e1);
+    s4 = deal(s2, e2);
+    if(s3 == s4 && e1 == e2) {
+        cout << "YES 0." << s3 << "*10^" << e1 << endl;
+    } else {
+        cout << "NO 0." << s3 << "*10^" << e1 << " 0." << s4 << "*10^" << e2 << endl;
     }
-    
     return 0;
-    }
+}
